Set up shared-brotli-patch-decoder to be able to utilize a rust or c brotli decoder implementation.
The pure rust decoder implementation is needed when compiling to WASM which can't easily work with the c version.
